Change ordering of cells and edit pf generator
Change ordering of cells to be linear. Change the bounds in the
portfolio generator function to accommodate the requirements.

Cells were ordered in a way such that functions were being called before
they were defined. Bounds in the portfolio generator were incorrect, as
individual stocks could take up anywhere from 0 to 100% of the
portfolio.