Fix indices of market_returns_df and ticker_returns
Fix the index of market_returns_df so that any days where only one
market is open is dropped. Change the index of ticker_returns so that it
matches the index of market_returns_df.

The previous code would not drop days where only one market was open,
leading to the dataframe having values when one market is open but the
other is closed.